backend_functions/buckets/upload_file/upload_finalize/requirements.txt
google-cloud-storage
Flask
functions-framework
